function_utils.py moved into tools
