+++ autogen/agentchat/contrib/swarm_agent.py
@@ -13,8 +13,8 @@ from typing import Any, Callable, Dict, List, Literal, Optional, Tuple, Union
 
 from pydantic import BaseModel
 
-from autogen.function_utils import get_function_schema
 from autogen.oai import OpenAIWrapper
+from autogen.tools import get_function_schema
 
 from ..agent import Agent
 from ..chat import ChatResult
