+++ autogen/tools/__init__.py
@@ -3,6 +3,15 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from .dependency_injection import BaseContext, ChatContext, Depends
+from .function_utils import get_function_schema, load_basemodels_if_needed, serialize_to_str
 from .tool import Tool
 
-__all__ = ["BaseContext", "ChatContext", "Depends", "Tool"]
+__all__ = [
+    "BaseContext",
+    "ChatContext",
+    "Depends",
+    "get_function_schema",
+    "load_basemodels_if_needed",
+    "serialize_to_str",
+    "Tool",
+]
