+++ pyproject.toml
@@ -58,13 +58,11 @@ files = [
     "autogen/coding",
     "autogen/oai/openai_utils.py",
     "autogen/_pydantic.py",
-    "autogen/function_utils.py",
     "autogen/io",
     "autogen/tools",
     "autogen/interop",
     "autogen/agentchat/realtime_agent",
     "test/test_pydantic.py",
-    "test/test_function_utils.py",
     "test/io",
     "test/tools",
     "test/interop",
