+++ website/docs/topics/openai-assistant/gpt_assistant_agent.mdx
@@ -146,7 +146,7 @@ Using the Function calling with the following configuration.
 
 ```python
 # learn more from https://platform.openai.com/docs/guides/function-calling/function-calling
-from autogen.function_utils import get_function_schema
+from autogen.tools import get_function_schema
 
 def get_current_weather(location: str) -> dict:
     """
